moved permissions checks to avoid cases where they could be bypassed; code cleanup
