+++ examples/pxScene2d/src/pxImage.cpp
@@ -125,7 +125,11 @@ rtError pxImage::url(rtString& s) const
   return RT_OK;
 }
 rtError pxImage::setUrl(const char* s)
-{ 
+{
+#ifdef ENABLE_PERMISSIONS_CHECK
+  rtPermissionsCheck((mScene != NULL ? mScene->permissions() : NULL), s, rtPermissions::DEFAULT)
+#endif
+
   //rtLogInfo("pxImage::setUrl init=%d imageLoaded=%d \n", mInitialized, imageLoaded);
   //rtLogDebug("pxImage::setUrl for s=%s mUrl=%s\n", s, mUrl.cString());
