+++ examples/pxScene2d/src/pxScene2d.cpp
@@ -1848,10 +1848,12 @@ pxScene2d::pxScene2d(bool top, pxScriptView* scriptView)
   if (scriptView != NULL)
   {
     mOrigin = rtUrlGetOrigin(scriptView->getUrl().cString());
-#ifdef ENABLE_PERMISSIONS_CHECK
-    mPermissions.setOrigin(mOrigin);
-#endif
   }
+
+#ifdef ENABLE_PERMISSIONS_CHECK
+  // rtPermissions accounts parent scene permissions too
+  mPermissions = new rtPermissions(mOrigin.cString());
+#endif
 
   // make sure that initial onFocus is sent
   rtObjectRef e = new rtMapObject;
@@ -1956,15 +1958,6 @@ void pxScene2d::init()
 
 rtError pxScene2d::create(rtObjectRef p, rtObjectRef& o)
 {
-#ifdef ENABLE_PERMISSIONS_CHECK
-  rtString url = p.get<rtString>("url");
-  if (!mPermissions.allows(url.cString(), rtPermissions::DEFAULT))
-  {
-    rtLogError("url '%s' is not allowed", url.cString());
-    return RT_ERROR_NOT_ALLOWED;
-  }
-#endif
-
   rtError e = RT_OK;
   rtString t = p.get<rtString>("t");
   bool needpxObjectTracking = true;
@@ -2126,6 +2119,11 @@ rtError pxScene2d::createImageResource(rtObjectRef p, rtObjectRef& o)
 {
   rtString url = p.get<rtString>("url");
   rtString proxy = p.get<rtString>("proxy");
+
+#ifdef ENABLE_PERMISSIONS_CHECK
+  rtPermissionsCheck(mPermissions, url.cString(), rtPermissions::DEFAULT)
+#endif
+
   o = pxImageManager::getImage(url, proxy);
   o.send("init");
   return RT_OK;
@@ -2135,6 +2133,11 @@ rtError pxScene2d::createImageAResource(rtObjectRef p, rtObjectRef& o)
 {
   rtString url = p.get<rtString>("url");
   rtString proxy = p.get<rtString>("proxy");
+
+#ifdef ENABLE_PERMISSIONS_CHECK
+  rtPermissionsCheck(mPermissions, url.cString(), rtPermissions::DEFAULT)
+#endif
+
   o = pxImageManager::getImageA(url, proxy);
   o.send("init");
   return RT_OK;
@@ -2144,6 +2147,11 @@ rtError pxScene2d::createFontResource(rtObjectRef p, rtObjectRef& o)
 {
   rtString url = p.get<rtString>("url");
   rtString proxy = p.get<rtString>("proxy");
+
+#ifdef ENABLE_PERMISSIONS_CHECK
+  rtPermissionsCheck(mPermissions, url.cString(), rtPermissions::DEFAULT)
+#endif
+
   o = pxFontManager::getFont(url, proxy);
   return RT_OK;
 }
@@ -2151,9 +2159,6 @@ rtError pxScene2d::createFontResource(rtObjectRef p, rtObjectRef& o)
 rtError pxScene2d::createScene(rtObjectRef p, rtObjectRef& o)
 {
   pxSceneContainer* sceneContainer = new pxSceneContainer(this);
-#ifdef ENABLE_PERMISSIONS_CHECK
-  sceneContainer->setParentPermissions(&mPermissions);
-#endif
   o = sceneContainer;
   o.set(p);
   o.send("init");
@@ -2195,15 +2200,6 @@ rtError pxScene2d::createExternal(rtObjectRef p, rtObjectRef& o)
 
 rtError pxScene2d::createWayland(rtObjectRef p, rtObjectRef& o)
 {
-#ifdef ENABLE_PERMISSIONS_CHECK
-  rtString cmd = p.get<rtString>("cmd");
-  if (!mPermissions.allows(cmd.cString(), rtPermissions::WAYLAND))
-  {
-    rtLogError("wayland cmd '%s' is not allowed", cmd.cString());
-    return RT_ERROR_NOT_ALLOWED;
-  }
-#endif
-
 #if defined(ENABLE_DFB) || defined(DISABLE_WAYLAND)
   UNUSED_PARAM(p);
   UNUSED_PARAM(o);
@@ -3045,11 +3041,7 @@ rtError pxScene2d::setCustomAnimator(const rtFunctionRef& v)
 rtError pxScene2d::screenshot(rtString type, rtString& pngData)
 {
 #ifdef ENABLE_PERMISSIONS_CHECK
-  if (!mPermissions.allows("screenshot", rtPermissions::FEATURE))
-  {
-    rtLogError("screenshot is not allowed");
-    return RT_ERROR_NOT_ALLOWED;
-  }
+  rtPermissionsCheck(mPermissions, "screenshot", rtPermissions::FEATURE)
 #endif
 
   // Is this a type we support?
@@ -3119,6 +3111,10 @@ rtError pxScene2d::getService(rtString name, rtObjectRef& returnObject)
 {
   rtLogDebug("inside getService");
   returnObject = NULL;
+
+#ifdef ENABLE_PERMISSIONS_CHECK
+  rtPermissionsCheck(mPermissions, name.cString(), rtPermissions::SERVICE)
+#endif
 
   // Create context from requesting scene
   rtObjectRef ctx = new rtMapObject();
@@ -3208,14 +3204,6 @@ rtError pxScene2d::getService(const char* name, const rtObjectRef& ctx, rtObject
   {
     // TODO JRJR should move this to top level container only...
 
-  #ifdef ENABLE_PERMISSIONS_CHECK
-    if (!mPermissions.allows(name, rtPermissions::SERVICE))
-    {
-      rtLogError("service '%s' is not allowed", name);
-      return RT_ERROR_NOT_ALLOWED;
-    }
-  #endif
-
     rtLogInfo("trying to get service for name: %s", name);
   #ifdef PX_SERVICE_MANAGER
     rtObjectRef serviceManager;
@@ -3290,6 +3278,9 @@ rtDefineProperty(pxScene2d,truncation);
 rtDefineMethod(pxScene2d, dispose);
 
 rtDefineProperty(pxScene2d, origin);
+#ifdef ENABLE_PERMISSIONS_CHECK
+rtDefineProperty(pxScene2d, permissions);
+#endif
 rtDefineMethod(pxScene2d, allows);
 rtDefineMethod(pxScene2d, checkAccessControlHeaders);
 rtDefineMethod(pxScene2d, addServiceProvider);
@@ -3401,7 +3392,7 @@ void pxScene2d::innerpxObjectDisposed(rtObjectRef ref)
 rtError pxScene2d::allows(const rtString& url, bool& o) const
 {
 #ifdef ENABLE_PERMISSIONS_CHECK
-  return mPermissions.allows(url.cString(), rtPermissions::DEFAULT, o);
+  return mPermissions->allows(url.cString(), rtPermissions::DEFAULT, o);
 #else
   UNUSED_PARAM(url);
   o = true; // default
@@ -3419,6 +3410,19 @@ rtError pxScene2d::checkAccessControlHeaders(const rtString& url, const rtString
   UNUSED_PARAM(rawHeaders);
   allow = true; // default
   return RT_OK;
+#endif
+}
+
+void pxScene2d::setViewContainer(pxIViewContainer* l)
+{
+  mContainer = l;
+#ifdef ENABLE_PERMISSIONS_CHECK
+  pxObject* obj = dynamic_cast<pxObject*>(l);
+  if (obj != NULL && obj->getScene())
+  {
+    // rtPermissions accounts parent scene permissions too
+    mPermissions->setParent(obj->getScene()->mPermissions);
+  }
 #endif
 }
 
@@ -3449,6 +3453,11 @@ rtDefineProperty(pxSceneContainer, ready);
 rtError pxSceneContainer::setUrl(rtString url)
 {
   rtLogInfo("pxSceneContainer::setUrl(%s)",url.cString());
+
+#ifdef ENABLE_PERMISSIONS_CHECK
+  rtPermissionsCheck((mScene != NULL ? mScene->permissions() : NULL), url.cString(), rtPermissions::DEFAULT)
+#endif
+
   // If old promise is still unfulfilled resolve it
   // and create a new promise for the context of this Url
   mReady.send("resolve", this);
@@ -3499,26 +3508,6 @@ rtError pxSceneContainer::setScriptView(pxScriptView* scriptView)
   return RT_OK;
 }
 
-#ifdef ENABLE_PERMISSIONS_CHECK
-rtError pxSceneContainer::setParentPermissions(const rtPermissions* v)
-{
-  if (mScriptView)
-  {
-    return mScriptView->setParentPermissions(v);
-  }
-  return RT_FAIL;
-}
-
-rtError pxSceneContainer::setPermissions(const rtObjectRef& v)
-{
-  if (mScriptView)
-  {
-    return mScriptView->setPermissions(v);
-  }
-  return RT_FAIL;
-}
-#endif
-
 void pxSceneContainer::dispose()
 {
   if (!mIsDisposed)
@@ -3540,6 +3529,13 @@ void pxSceneContainer::dispose()
     }
     return NULL;
   }
+
+#ifdef ENABLE_PERMISSIONS_CHECK
+rtError pxSceneContainer::setPermissions(const rtObjectRef& v)
+{
+  return mScriptView != NULL ? mScriptView->setPermissions(v) : RT_FAIL;
+}
+#endif
 
 #if 0
 void* gObjectFactoryContext = NULL;
@@ -3746,25 +3742,3 @@ rtError pxScriptView::makeReady(int numArgs, const rtValue* args, rtValue* /*res
   }
   return RT_FAIL;
 }
-
-#ifdef ENABLE_PERMISSIONS_CHECK
-rtError pxScriptView::setParentPermissions(const rtPermissions* v)
-{
-  if (mScene)
-  {
-    pxScene2d* s = (pxScene2d*)mScene.getPtr();
-    return s->setParentPermissions(v);
-  }
-  return RT_FAIL;
-}
-
-rtError pxScriptView::setPermissions(const rtObjectRef& v)
-{
-  if (mScene)
-  {
-    pxScene2d* s = (pxScene2d*)mScene.getPtr();
-    return s->setPermissions(v);
-  }
-  return RT_FAIL;
-}
-#endif
